requirements.txt
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
sentence-transformers>=2.2.2
faiss-gpu>=1.7.4
vllm>=0.2.6
numpy
tqdm
